Use NodeUtils for string Nodes
